ecommerce/requirements.txt
asgiref==3.11.0
attrs==25.4.0
colorama==0.4.6
Django==6.0
django-filter==25.2
django-rest-framework==0.1.0
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gprof2dot==2025.4.14
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
